+++ packages/terra-clinical-application-site/src/demo/patient-detail/PatientDetailController.jsx
@@ -4,8 +4,8 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import AppDelegate from 'terra-clinical-app-delegate';
 
-import { loadPatient } from '../patient-concept/actions/patientActions';
-import patientReducers from '../patient-concept/reducers/patientReducers';
+import { loadPatient } from '../patient-concept/actions';
+import patientReducers from '../patient-concept/reducers';
 
 import PatientDetail from './PatientDetail';
